module parser moved to a separate file from main.rs
